pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hei-seti"
version = "0.1.0"
description = "High-Energy Astrobiology toolkit for Kardashev-Barrow anomaly searches"
readme = "README.md"
authors = [{name = "Lucky + Monday"}]
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
  "pydantic>=2.7",
  "astropy>=6.0",
  "astroquery>=0.4.7",
  "lightkurve>=2.4; python_version < '3.13'",
  "pyarrow>=17.0.0",
  "tqdm>=4.66",
  "joblib>=1.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.4",
  "black>=24.3"
]

[project.scripts]
hei-seti = "hei_seti.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra --cov=hei_seti --cov-report=term-missing"
minversion = "8.0"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']
